Moved the data files to a dedicated folder
